refactor: extract voice playback hook
